tests/wpt/tests/css/css-multicol/animation/discrete-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'column-fill',
  from: 'initial',
  to: 'auto'
});

test_no_interpolation({
  property: 'column-rule-style',
  from: 'initial',
  to: 'dashed'
});

test_no_interpolation({
  property: 'column-span',
  from: 'initial',
  to: 'all'
});
</script>
